Move vision detection to filter
