+++ src/software/proto/message_translation/ssl_detection.h
@@ -4,7 +4,7 @@
 #include <memory>
 
 #include "software/proto/messages_robocup_ssl_detection.pb.h"
-#include "software/sensor_fusion/vision_detection.h"
+#include "software/sensor_fusion/filter/vision_detection.h"
 #include "software/time/timestamp.h"
 #include "software/world/ball_state.h"
 #include "software/world/robot_state.h"
