+++ src/software/sensor_fusion/BUILD
@@ -1,18 +1,6 @@
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "vision_detection",
-    hdrs = [
-        "vision_detection.h",
-    ],
-    deps = [
-        "//software/new_geom:angle",
-        "//software/new_geom:point",
-        "//software/time:timestamp",
-    ],
-)
-
-cc_library(
     name = "refbox_data",
     hdrs = ["refbox_data.h"],
     deps = [
@@ -35,7 +23,6 @@ cc_library(
     hdrs = ["sensor_fusion.h"],
     deps = [
         ":refbox_data",
-        ":vision_detection",
         "//software/backend:robot_status",
         "//software/multithreading:subject",
         "//software/multithreading:threaded_observer",
@@ -46,6 +33,7 @@ cc_library(
         "//software/proto/message_translation:ssl_geometry",
         "//software/proto/message_translation:ssl_referee",
         "//software/sensor_fusion/filter:sensor_fusion_filters",
+        "//software/sensor_fusion/filter:vision_detection",
         "//software/world",
     ],
 )
