+++ src/software/sensor_fusion/sensor_fusion.h
@@ -11,8 +11,8 @@
 #include "software/proto/sensor_msg.pb.h"
 #include "software/sensor_fusion/filter/ball_filter.h"
 #include "software/sensor_fusion/filter/robot_team_filter.h"
+#include "software/sensor_fusion/filter/vision_detection.h"
 #include "software/sensor_fusion/refbox_data.h"
-#include "software/sensor_fusion/vision_detection.h"
 #include "software/world/ball.h"
 #include "software/world/team.h"
 #include "software/world/world.h"
